Add login.sh to semi-automate the Bitwarden login step
login.sh is pretty basic currently. It logins to the bitwarden accounts using the credentials fetched.
(Bitwarden required OTP verification on new devices. This logic is not handled yet.) It then set the session cookies
and exports the vault.

This commit also removes a redundant folder "testing/" and modifies the server scripts slightly